+++ example/android/app/src/main/kotlin/com/stacked/firebaseexample/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.stacked.firebaseexample
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}