+++ lib/src/firebase_authentication_service.dart
@@ -4,6 +4,7 @@ import 'dart:math';
 import 'package:crypto/crypto.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/foundation.dart';
+import 'package:flutter_facebook_auth/flutter_facebook_auth.dart';
 import 'package:google_sign_in/google_sign_in.dart';
 import 'package:logger/logger.dart';
 import 'package:sign_in_with_apple/sign_in_with_apple.dart';
@@ -16,13 +17,15 @@ class FirebaseAuthenticationService {
   final Logger? log;
 
   final firebaseAuth = FirebaseAuth.instance;
-  final _googleSignIn = GoogleSignIn();
+  final GoogleSignIn _googleSignIn = GoogleSignIn();
 
   FirebaseAuthenticationService({
-    @Deprecated('Pass in the appleRedirectUri through the signInWithApple function')
-        String? appleRedirectUri,
-    @Deprecated('Pass in the appleClientId through the signInWithApple function')
-        String? appleClientId,
+    @Deprecated(
+        'Pass in the appleRedirectUri through the signInWithApple function')
+    String? appleRedirectUri,
+    @Deprecated(
+        'Pass in the appleClientId through the signInWithApple function')
+    String? appleClientId,
     this.log,
   });
 
@@ -56,6 +59,10 @@ class FirebaseAuthenticationService {
   }
 
   /// Returns `true` when email has a user registered
+  @Deprecated(
+    'emailExists() has been deprecated. '
+    'Migrating off of this method is recommended as a security best-practice. Learn more in the Identity Platform documentation: ',
+  )
   Future<bool> emailExists(String email) async {
     try {
       final signInMethods =
@@ -73,7 +80,8 @@ class FirebaseAuthenticationService {
   ///   - Android
   ///   - iOS
   ///   - Web
-  Future<FirebaseAuthenticationResult> signInWithGoogle() async {
+  Future<FirebaseAuthenticationResult> signInWithGoogle(
+      {String? webLoginHint}) async {
     try {
       UserCredential userCredential;
 
@@ -81,7 +89,8 @@ class FirebaseAuthenticationService {
       /// handling the authentication flow.
       if (kIsWeb) {
         GoogleAuthProvider googleProvider = GoogleAuthProvider();
-        googleProvider.setCustomParameters({'login_hint': 'user@example.com'});
+        googleProvider.setCustomParameters(
+            {'login_hint': webLoginHint ?? 'user@example.com'});
 
         userCredential = await FirebaseAuth.instance.signInWithPopup(
           googleProvider,
@@ -109,6 +118,78 @@ class FirebaseAuthenticationService {
         );
 
         userCredential = await _signInWithCredential(credential);
+      }
+
+      // Link the pending credential with the existing account
+      if (_pendingCredential != null) {
+        await userCredential.user?.linkWithCredential(_pendingCredential!);
+        _clearPendingData();
+      }
+
+      return FirebaseAuthenticationResult(
+        user: userCredential.user,
+        additionalUserInfo: userCredential.additionalUserInfo,
+      );
+    } on FirebaseAuthException catch (e) {
+      log?.e(e);
+      return FirebaseAuthenticationResult.error(
+        errorMessage: getErrorMessageFromFirebaseException(e),
+        exceptionCode: e.code,
+      );
+    } catch (e) {
+      log?.e(e);
+      return FirebaseAuthenticationResult.error(errorMessage: e.toString());
+    }
+  }
+
+  /// Authenticates a user through Firebase using Facebook Provider.
+  ///
+  /// Supported platforms:
+  ///   - Android
+  ///   - iOS
+  ///   - Web
+  Future<FirebaseAuthenticationResult> signInWithFacebook(
+      {String? webLoginHint}) async {
+    try {
+      UserCredential userCredential;
+
+      /// On the web, the Firebase SDK provides support for automatically
+      /// handling the authentication flow.
+      if (kIsWeb) {
+        FacebookAuthProvider facebookProvider = FacebookAuthProvider();
+        facebookProvider.setCustomParameters(
+            {'login_hint': webLoginHint ?? 'user@example.com'});
+
+        userCredential = await FirebaseAuth.instance.signInWithPopup(
+          facebookProvider,
+        );
+      }
+
+      /// On native platforms, a 3rd party library, like FacebookSignIn, is
+      /// required to trigger the authentication flow.
+      else {
+        final LoginResult facebookLoginResult =
+            await FacebookAuth.instance.login();
+        if (facebookLoginResult.status == LoginStatus.cancelled) {
+          log?.i('Process is canceled by the user');
+          return FirebaseAuthenticationResult.error(
+            errorMessage: 'Facebook Sign In has been canceled by the user',
+            exceptionCode: 'canceled',
+          );
+        } else if (facebookLoginResult.status == LoginStatus.failed) {
+          log?.i('Login failed with error: ${facebookLoginResult.message}');
+          return FirebaseAuthenticationResult.error(
+            errorMessage:
+                'Facebook Sign In has failed with error: ${facebookLoginResult.message}',
+            exceptionCode: 'failed',
+          );
+        }
+
+        final OAuthCredential facebookAuthCredential =
+            FacebookAuthProvider.credential(
+                facebookLoginResult.accessToken!.token);
+
+        userCredential = await _signInWithCredential(facebookAuthCredential);
       }
 
       // Link the pending credential with the existing account
@@ -230,17 +311,17 @@ class FirebaseAuthenticationService {
   /// Anonymous Login
   Future<FirebaseAuthenticationResult> loginAnonymously() async {
     try {
-      log?.d('Anonymoys Login');
+      log?.d('Anonymous Login');
       final result = await firebaseAuth.signInAnonymously();
 
       return FirebaseAuthenticationResult(user: result.user);
     } on FirebaseAuthException catch (e) {
-      log?.e('A firebase exception has occured. $e');
+      log?.e('A firebase exception has occurred. $e');
       return FirebaseAuthenticationResult.error(
           exceptionCode: e.code.toLowerCase(),
           errorMessage: getErrorMessageFromFirebaseException(e));
     } on Exception catch (e) {
-      log?.e('A general exception has occured. $e');
+      log?.e('A general exception has occurred. $e');
       return FirebaseAuthenticationResult.error(
           errorMessage:
               'We could not log into your account at this time. Please try again.');
@@ -268,12 +349,12 @@ class FirebaseAuthenticationService {
 
       return FirebaseAuthenticationResult(user: result.user);
     } on FirebaseAuthException catch (e) {
-      log?.e('A firebase exception has occured. $e');
+      log?.e('A firebase exception has occurred. $e');
       return FirebaseAuthenticationResult.error(
           exceptionCode: e.code.toLowerCase(),
           errorMessage: getErrorMessageFromFirebaseException(e));
     } on Exception catch (e) {
-      log?.e('A general exception has occured. $e');
+      log?.e('A general exception has occurred. $e');
       return FirebaseAuthenticationResult.error(
           errorMessage:
               'We could not log into your account at this time. Please try again.');
@@ -297,12 +378,12 @@ class FirebaseAuthenticationService {
 
       return FirebaseAuthenticationResult(user: result.user);
     } on FirebaseAuthException catch (e) {
-      log?.e('A firebase exception has occured. $e');
+      log?.e('A firebase exception has occurred. $e');
       return FirebaseAuthenticationResult.error(
           exceptionCode: e.code.toLowerCase(),
           errorMessage: getErrorMessageFromFirebaseException(e));
     } on Exception catch (e) {
-      log?.e('A general exception has occured. $e');
+      log?.e('A general exception has occurred. $e');
       return FirebaseAuthenticationResult.error(
           errorMessage:
               'We could not create your account at this time. Please try again.');
@@ -412,13 +493,13 @@ class FirebaseAuthenticationService {
 
       return FirebaseAuthenticationResult(user: userCredential.user);
     } on FirebaseAuthException catch (e) {
-      log?.e('A Firebase exception has occured. $e');
+      log?.e('A Firebase exception has occurred. $e');
       return FirebaseAuthenticationResult.error(
         exceptionCode: e.code.toLowerCase(),
         errorMessage: getErrorMessageFromFirebaseException(e),
       );
     } on Exception catch (e) {
-      log?.e('A general exception has occured. $e');
+      log?.e('A general exception has occurred. $e');
       return FirebaseAuthenticationResult.error(
         errorMessage:
             'We could not authenticate with OTP at this time. Please try again.',
@@ -426,14 +507,75 @@ class FirebaseAuthenticationService {
     }
   }
 
-  /// Sign out of the social accounts that have been used
-  Future logout() async {
-    log?.i('');
+  /// Validate phone number using the [otp] and link it with user account
+  ///
+  /// A [FirebaseAuthenticationResult] maybe returned with the following error
+  /// code:
+  /// - **mobileVerificationId-not-present**:
+  ///  - Returned if the [_mobileVerificationId] is null. This error could
+  ///    happen if the verification was skipped.
+  /// - **user-not-logged-in**:
+  ///  - Returned if the user is not logged in the Firebase instance.
+  /// - **link-phone-number-failure**:
+  ///  - Returned if the phone number could NOT be linked into the account.
+  Future<FirebaseAuthenticationResult> validateOtpAndLinkPhoneNumber(
+    String otp,
+  ) async {
+    if (_mobileVerificationId == null) {
+      return FirebaseAuthenticationResult.error(
+        errorMessage:
+            'The _mobileVerificationId should not be null here. Verification was probably skipped.',
+        exceptionCode: 'mobileVerificationId-not-present',
+      );
+    }
+
+    if (!hasUser) {
+      return FirebaseAuthenticationResult.error(
+        errorMessage:
+            'The Firebase instance has no User which means the user is not logged in, please sign in the user before link the phone number.',
+        exceptionCode: 'user-not-logged-in',
+      );
+    }
 
     try {
+      final phoneAuthCredential = PhoneAuthProvider.credential(
+        verificationId: _mobileVerificationId!,
+        smsCode: otp,
+      );
+
+      final userCredential = await firebaseAuth.currentUser!.linkWithCredential(
+        phoneAuthCredential,
+      );
+
+      return FirebaseAuthenticationResult(
+        user: userCredential.user,
+        additionalUserInfo: userCredential.additionalUserInfo,
+      );
+    } on FirebaseAuthException catch (e) {
+      log?.e('A Firebase exception has occurred. $e');
+      return FirebaseAuthenticationResult.error(
+        exceptionCode: e.code.toLowerCase(),
+        errorMessage: getErrorMessageFromFirebaseException(e),
+      );
+    } on Exception catch (e) {
+      log?.e('A general exception has occurred. $e');
+      return FirebaseAuthenticationResult.error(
+        errorMessage:
+            'We could not link your phone number at this time. Please try again.',
+        exceptionCode: 'link-phone-number-failure',
+      );
+    }
+  }
+
+  /// Sign out of the social accounts that have been used
+  Future<void> logout() async {
+    log?.i('logout');
+
+    try {
+      _clearPendingData();
       await firebaseAuth.signOut();
       await _googleSignIn.signOut();
-      _clearPendingData();
+      await FacebookAuth.instance.logOut();
     } catch (e) {
       log?.e('Could not sign out of social account. $e');
     }
@@ -444,7 +586,7 @@ class FirebaseAuthenticationService {
   }
 
   /// Send reset password link to email
-  Future sendResetPasswordLink(String email) async {
+  Future<bool> sendResetPasswordLink(String email) async {
     log?.i('email:$email');
 
     try {
@@ -457,7 +599,7 @@ class FirebaseAuthenticationService {
   }
 
   /// Validate the current [password] of the Firebase User
-  Future validatePassword(String password) async {
+  Future<dynamic> validatePassword(String password) async {
     try {
       final authCredentials = EmailAuthProvider.credential(
         email: firebaseAuth.currentUser?.email ?? '',
@@ -471,18 +613,37 @@ class FirebaseAuthenticationService {
     } catch (e) {
       log?.e('Could not validate the user password. $e');
       return FirebaseAuthenticationResult.error(
-          errorMessage: 'The current password is not valid.');
+        errorMessage: 'The current password is not valid.',
+      );
     }
   }
 
   /// Update the [password] of the Firebase User
-  Future updatePassword(String password) async {
+  Future<void> updatePassword(String password) async {
     await firebaseAuth.currentUser?.updatePassword(password);
   }
 
   /// Update the [email] of the Firebase User
-  Future updateEmail(String email) async {
+  @Deprecated(
+    'updateEmail() has been deprecated. Please use verifyBeforeUpdateEmail() instead.',
+  )
+  Future<void> updateEmail(String email) async {
     await firebaseAuth.currentUser?.updateEmail(email);
+  }
+
+  /// Sends a verification email to a new email address. The user's email will be updated to the new one after being verified.
+  Future<void> verifyBeforeUpdateEmail(String email) async {
+    await firebaseAuth.currentUser?.verifyBeforeUpdateEmail(email);
+  }
+
+  /// Update the [displayName] of the Firebase User
+  Future<void> updateDisplayName(String displayName) async {
+    await firebaseAuth.currentUser?.updateDisplayName(displayName);
+  }
+
+  /// Update the [photoURL] of the Firebase User
+  Future<void> updatePhotoURL(String photoUrl) async {
+    await firebaseAuth.currentUser?.updatePhotoURL(photoUrl);
   }
 
   /// Generates a cryptographically secure random nonce, to be included in a